requirements.txt
django 
django-tinymce4-lite
django-filebrowser-no-grappelli
pillow 
django-rest-framework
django-import-export 
